Updated aoa_sensor example which now listens for a button push on a GPIO pin to update the EEPROM with whatever raw angle the rotary encoder/sensor is reading at the moment. This should be done when flying straight and level.
Also, before the CAN-FIX message is sent, that zeroizing value is subtracted from the reading sent down the wire.

Eventually, the Arduino will listen for a custom CAN id parameter (probably sent from pyEfis?) as a signal to reset one of its values
to whatever the raw value on the rotary encoder is reporting at the time the button is pressed. This should be done for each of the AOA
parameters (0g, warning, & stall)