Added a new AoA example using the MPXV7007DP pressure differential sensor (rather than a rotary encoder).
This sensor has two pressure lines. In this instance, both are connected to a pitot tube with one of the holes offset by some amount.
The sensor will read the pressure difference between the two tubes. As the angle of attack increases, the pressure on the offset hole
will increase proportionally to the top inlet. Using the aoa_sensor_angle_setter (another Arduino project), you can press a button and
set the 0g angle (this will send a specific CAN message that this project listens for). Whatever the sensor reads when you press the
button will be written to this Arduino's EEPROM and should not need to be reset again. You can do the same to set the Stall angle when
the aircraft is approaching a stall. The stall value will also be written to the Arduino's EEPROM.

This project will then, every second, push it's parameters (e.g. 0g and stall values) out to the CAN Bus where CAN-FiX will publish them to pyEfis.
My branch of pyEfis has an AoA meter on the PFD screen which interpolates the difference between the 0g and the Stall values and will light up
bars on the screen as AoA increases. It will additionally play a soundbyte that says "nose down" once the Stall value has been reached.